Clean reports, Relocated factory, Querying Teams
